Move helper functions to the same file
