+++ src/components/Contacts/ContactsMap/ContactsMapPanel.tsx
@@ -21,7 +21,7 @@ import { useTranslation } from 'react-i18next';
 import { PhaseEnum, StatusEnum } from 'src/graphql/types.generated';
 import { useContactPartnershipStatuses } from 'src/hooks/useContactPartnershipStatuses';
 import theme from 'src/theme';
-import { sourceToStr } from 'src/utils/sourceToStr';
+import { sourceToStr } from 'src/utils/sourceHelper';
 import { Coordinates } from './coordinates';
 
 interface ContactMapsPanelProps {
