--- src/components/Tool/helpers.test.ts
@@ -1,8 +1,0 @@
-import { isEditableSource } from './helpers';
-
-describe('Tool-Helpers', () => {
-  it('isEditableSource', () => {
-    expect(isEditableSource('MPDX')).toBe(true);
-    expect(isEditableSource('Siebel')).toBe(false);
-  });
-});
