--- src/components/Tool/helpers.ts
@@ -1,4 +1,0 @@
-export const editableSources = ['MPDX', 'manual', 'TntImport'];
-
-export const isEditableSource = (source: string) =>
-  editableSources.indexOf(source) > -1;
